refactor!: remove the remote AI worker protocol
The standalone -aiWorker node only ever dispatched batch pipelines,
so with those gone the AIWorker gRPC service, the /aiResults endpoint
and RemoteAIWorkerManager are dead. -aiWorker now exits unless
combined with -orchestrator, and the orchestrator serves
live-video-to-video from its local worker only.

BREAKING CHANGE: the standalone AI worker node type and the
net.AIWorker gRPC service are removed.
