+++ cmd/livepeer/starter/starter.go
@@ -75,7 +75,6 @@ const (
 	OrchestratorRpcPort = "8935"
 	OrchestratorCliPort = "7935"
 	TranscoderCliPort   = "6935"
-	AIWorkerCliPort     = "4935"
 	RemoteSignerCliPort = "3935"
 
 	RefreshPerfScoreInterval = 10 * time.Minute
@@ -738,13 +737,10 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 			n.TranscoderManager = core.NewRemoteTranscoderManager()
 			n.Transcoder = n.TranscoderManager
 		}
-		if !*cfg.AIWorker {
-			n.AIWorkerManager = core.NewRemoteAIWorkerManager()
-		}
 	} else if *cfg.Transcoder {
 		n.NodeType = core.TranscoderNode
 	} else if *cfg.AIWorker {
-		n.NodeType = core.AIWorkerNode
+		glog.Exit("-aiWorker requires -orchestrator: the standalone AI worker node type was removed together with the batch pipelines")
 	} else if *cfg.Broadcaster {
 		n.NodeType = core.BroadcasterNode
 		glog.Warning("-broadcaster flag is deprecated and will be removed in a future release. Please use -gateway instead")
@@ -778,8 +774,6 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 			nodeType = lpmon.Transcoder
 		case core.RedeemerNode:
 			nodeType = lpmon.Redeemer
-		case core.AIWorkerNode:
-			nodeType = lpmon.AIWorker
 		}
 		lpmon.InitCensus(nodeType, core.LivepeerVersion)
 	}
@@ -1563,11 +1557,9 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 				n.SetBasePriceForCap(config.Gateway, pipelineCap, config.ModelID, autoPrice)
 			}
 		}
-	} else {
-		if n.NodeType == core.AIWorkerNode {
-			glog.Error("The '-aiWorker' flag was set, but no model configuration was provided. Please specify the model configuration using the '-aiModels' flag.")
-			return
-		}
+	} else if *cfg.AIWorker {
+		glog.Error("The '-aiWorker' flag was set, but no model configuration was provided. Please specify the model configuration using the '-aiModels' flag.")
+		return
 	}
 
 	if *cfg.Objectstore != "" {
@@ -1812,8 +1804,6 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 		}
 	} else if n.NodeType == core.TranscoderNode {
 		*cfg.CliAddr = defaultAddr(*cfg.CliAddr, "127.0.0.1", TranscoderCliPort)
-	} else if n.NodeType == core.AIWorkerNode {
-		*cfg.CliAddr = defaultAddr(*cfg.CliAddr, "127.0.0.1", AIWorkerCliPort)
 	} else if n.NodeType == core.RemoteSignerNode {
 		*cfg.CliAddr = defaultAddr(*cfg.CliAddr, "127.0.0.1", RemoteSignerCliPort)
 	}
@@ -1822,7 +1812,7 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 	}
 
 	// Apply default capabilities if not running as a transcoder.
-	if !*cfg.Transcoder && (n.NodeType == core.AIWorkerNode || n.NodeType == core.OrchestratorNode) {
+	if !*cfg.Transcoder && n.NodeType == core.OrchestratorNode {
 		aiCaps = append(aiCaps, core.DefaultCapabilities()...)
 	}
 
@@ -1833,10 +1823,6 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 	}
 	if cfg.AIMinRunnerVersion != nil {
 		n.Capabilities.SetMinRunnerVersionConstraint(*cfg.AIMinRunnerVersion)
-	}
-	if n.AIWorkerManager != nil {
-		// Set min version constraint to prevent incompatible workers.
-		n.Capabilities.SetMinVersionConstraint(core.LivepeerVersion)
 	}
 
 	if drivers.NodeStorage == nil {
@@ -2058,7 +2044,7 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 		}
 
 		go func() {
-			err = server.StartTranscodeServer(orch, *cfg.HttpAddr, s.HTTPMux, n.WorkDir, n.TranscoderManager != nil, n.AIWorkerManager != nil, n)
+			err = server.StartTranscodeServer(orch, *cfg.HttpAddr, s.HTTPMux, n.WorkDir, n.TranscoderManager != nil, n)
 			if err != nil {
 				exit("Error starting Transcoder node: err=%q", err)
 			}
@@ -2094,21 +2080,14 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 
 	}()
 
-	if n.NodeType == core.TranscoderNode || n.NodeType == core.AIWorkerNode {
+	if n.NodeType == core.TranscoderNode {
 		if n.OrchSecret == "" {
 			glog.Exit("Missing -orchSecret")
 		}
 		if len(orchURLs) <= 0 {
 			glog.Exit("Missing -orchAddr")
 		}
-
-		if n.NodeType == core.TranscoderNode {
-			go server.RunTranscoder(n, orchURLs[0].Host, core.MaxSessions, transcoderCaps)
-		}
-
-		if n.NodeType == core.AIWorkerNode {
-			go server.RunAIWorker(n, orchURLs[0].Host, n.Capabilities.ToNetCapabilities())
-		}
+		go server.RunTranscoder(n, orchURLs[0].Host, core.MaxSessions, transcoderCaps)
 	}
 
 	switch n.NodeType {
@@ -2119,8 +2098,6 @@ func StartLivepeer(ctx context.Context, cfg LivepeerConfig) {
 		glog.Infof("Video Ingest Endpoint - rtmp://%v", *cfg.RtmpAddr)
 	case core.TranscoderNode:
 		glog.Infof("**Liveepeer Running in Transcoder Mode***")
-	case core.AIWorkerNode:
-		glog.Infof("**Livepeer Running in AI Worker Mode**")
 	case core.RedeemerNode:
 		glog.Infof("**Livepeer Running in Redeemer Mode**")
 	}
