+++ net/lp_rpc.proto
@@ -13,13 +13,6 @@ service Orchestrator {
   rpc Ping(PingPong) returns (PingPong);
 }
 
-service AIWorker {
-
-  // Called by the aiworker to register to an orchestrator. The orchestrator
-  // notifies registered aiworkers of jobs as they come in.
-  rpc RegisterAIWorker(RegisterAIWorkerRequest) returns (stream NotifyAIJob);
-}
-
 service Transcoder {
 
   // Called by the transcoder to register to an orchestrator. The orchestrator
@@ -406,37 +399,6 @@ message NotifySegment {
 
   // Deprecated by segData. Transcoding configuration to use.
   reserved 33; // Formerly "repeated VideoProfile fullProfiles"
-}
-
-// Sent by the aiworker to register itself to the orchestrator.
-message RegisterAIWorkerRequest {
-
-  // Shared secret for auth
-  string secret = 1;
-
-  // AIWorker capabilities
-  Capabilities capabilities = 2;
-
-  // AIWorker hardware info
-  repeated HardwareInformation hardware = 3;
-}
-
-// Data included by the gateway when submitting a AI job.
-message AIJobData {
-  // pipeline to use for the job
-  string pipeline = 1;
-
-  // AI job request data
-  bytes requestData = 2;
-}
-
-// Sent by the orchestrator to the aiworker
-message NotifyAIJob {
-  // Configuration for the AI job
-  AIJobData AIJobData = 1;
-
-  // ID for this particular AI task.
-  int64 taskId = 2;
 }
 
 // Required parameters for probabilistic micropayment tickets
